RN-964: Split QrCodeScreen into multiple components
