+++ packages/meditrak-app/app/assessment/qrCodeScreen/index.js
@@ -1,0 +1,6 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+export { QrCodeScreen } from './QrCodeScreen';
